// backend/packages/services-svc/src/utils/enums/accesses.enum.ts
export enum AccessesEnum {
  TODOS_MENUS = 'TODOS_MENUS',
  ADMINISTRATIVO = 'ADMINISTRATIVO',
  CLIENTES = 'CLIENTES',
  PEDIDOS = 'PEDIDOS',
  ORCAMENTO_E_NEGOCIACOES = 'ORCAMENTO_E_NEGOCIACOES',
  VENDAS = 'VENDAS',
  NOTAS_FICAIS_E_AGENDAMENTOS = 'NOTAS_FICAIS_E_AGENDAMENTOS',
  CONTRATOS = 'CONTRATOS',
  CATALOGOS = 'CATALOGOS',
  VPC = 'VPC',
  FINANCEIRO = 'FINANCEIRO',
  BIBLIOTECA = 'BIBLIOTECA',
  DASHBOARD = 'DASHBOARD',
  MENU = 'MENU',
  CONTROLE_DE_USUARIO = 'CONTROLE_DE_USUARIO',
  CONTROLE_DE_PERFIL = 'CONTROLE_DE_PERFIL',
  MURAL_DE_RECADOS = 'MURAL_DE_RECADOS',
  IMAGEM_DO_LOGIN = 'IMAGEM_DO_LOGIN',
  PAINEL_DE_WORKFLOW = 'PAINEL_DE_WORKFLOW',
  RANKING_DO_CLIENTE = 'RANKING_DO_CLIENTE',
  COMPRADORES = 'COMPRADORES',
  FAN = 'FAN',
  CONTAS_A_RECEBER = 'CONTAS_A_RECEBER',
  CADASTRO_EMPRESAS = 'CADASTRO_EMPRESAS',
  RELATORIOS_VPC = 'RELATORIOS_VPC'
}